services/api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
chromadb==0.5.11
python-dotenv==1.0.1
orjson==3.10.7
tenacity==8.5.0
pyyaml==6.0.2
